+++ src/Stack/SourceMap.hs
@@ -262,7 +262,7 @@ loadProjectSnapshotCandidate ::
     -> Bool
     -> RIO env (SnapshotCandidate env)
 loadProjectSnapshotCandidate loc printWarnings buildHaddocks = do
-    snapshot <- fmap fst . loadAndCompleteSnapshot =<< completeSnapshotLocation loc
+    (snapshot, _, _) <- loadAndCompleteSnapshotRaw loc Map.empty Map.empty
     deps <- Map.traverseWithKey (snapToDepPackage False) (snapshotPackages snapshot)
     let wc = snapshotCompiler snapshot
     globals <- Map.map GlobalPackageVersion <$> globalsFromHints wc