+++ subs/pantry/src/Pantry/Types.hs
@@ -90,6 +90,7 @@ module Pantry.Types
   , RawSnapshotLocation (..)
   , SnapshotLocation (..)
   , toRawSL
+  , parseHackageText
   , parseRawSnapshotLocation
   , RawSnapshotLayer (..)
   , SnapshotLayer (..)
@@ -105,6 +106,7 @@ module Pantry.Types
   , cabalFileName
   , SnapshotCacheHash (..)
   , getGlobalHintsFile
+  , bsToBlobKey
   ) where
 
 import RIO
@@ -504,6 +506,16 @@ instance Display Repo where
     (if T.null subdir
       then mempty
       else " in subdirectory " <> display subdir)
+instance FromJSON Repo where
+    parseJSON =
+        withObject "Repo" $ \o -> do
+            repoSubdir <- o .: "subdir"
+            repoCommit <- o .: "commit"
+            (repoType, repoUrl) <-
+                (o .: "git" >>= \url -> pure (RepoGit, url)) <|>
+                (o .: "hg" >>= \url -> pure (RepoHg, url))
+            pure Repo {..}
+
 
 -- An unexported newtype wrapper to hang a 'FromJSON' instance off of. Contains
 -- a GitHub user and repo name separated by a forward slash, e.g. "foo/bar".
@@ -710,6 +722,32 @@ instance FromJSON PackageIdentifierRevision where
       Left e -> fail $ show e
       Right pir -> pure pir
 
+-- | Parse a hackage text.
+parseHackageText :: Text -> Either PantryException (PackageIdentifier, BlobKey)
+parseHackageText t = maybe (Left $ PackageIdentifierRevisionParseFail t) Right $ do
+  let (identT, cfiT) = T.break (== '@') t
+  PackageIdentifier name version <- parsePackageIdentifier $ T.unpack identT
+  (csha, csize) <-
+    case splitColon cfiT of
+      Just ("@sha256", shaSizeT) -> do
+        let (shaT, sizeT) = T.break (== ',') shaSizeT
+        sha <- either (const Nothing) Just $ SHA256.fromHexText shaT
+        msize <-
+          case T.stripPrefix "," sizeT of
+            Nothing -> Nothing
+            Just sizeT' ->
+              case decimal sizeT' of
+                Right (size', "") -> Just $ (sha, FileSize size')
+                _ -> Nothing
+        pure msize
+      _ -> Nothing
+  pure $ (PackageIdentifier name version, BlobKey csha csize)
+
+splitColon :: Text -> Maybe (Text, Text)
+splitColon t' =
+    let (x, y) = T.break (== ':') t'
+     in (x, ) <$> T.stripPrefix ":" y
+
 -- | Parse a 'PackageIdentifierRevision'
 --
 -- @since 0.1.0.0
@@ -737,10 +775,6 @@ parsePackageIdentifierRevision t = maybe (Left $ PackageIdentifierRevisionParseF
       Nothing -> pure CFILatest
       _ -> Nothing
   pure $ PackageIdentifierRevision name version cfi
-  where
-    splitColon t' =
-      let (x, y) = T.break (== ':') t'
-       in (x, ) <$> T.stripPrefix ":" y
 
 data Mismatch a = Mismatch
   { mismatchExpected :: !a
@@ -1377,6 +1411,18 @@ instance Display PackageMetadata where
     , "cabal file == " <> display (pmCabal pm)
     ]
 
+instance FromJSON PackageMetadata where
+    parseJSON =
+        withObject "PackageMetadata" $ \o -> do
+            pmCabal :: BlobKey  <- o .: "cabal-file"
+            pantryTree :: BlobKey <- o .: "pantry-tree"
+            CabalString pkgName  <- o .: "name"
+            CabalString pkgVersion <- o .: "version"
+            let pmTreeKey = TreeKey pantryTree
+                pmIdent = PackageIdentifier {..}
+            pure PackageMetadata {..}
+
+
 -- | Conver package metadata to its "raw" equivalent.
 --
 -- @since 0.1.0.0
@@ -1489,6 +1535,54 @@ rpmToPairs (RawPackageMetadata mname mversion mtree mcabal) = concat
   , maybe [] (\cabal -> ["cabal-file" .= cabal]) mcabal
   ]
 
+instance FromJSON (WithJSONWarnings (Unresolved PackageLocationImmutable)) where
+    parseJSON v = repoObject v <|> archiveObject v <|> hackageObject v <|> github v
+                  <|> fail ("Could not parse a UnresolvedPackageLocationImmutable from: " ++ show v)
+        where
+          repoObject :: Value -> Parser (WithJSONWarnings (Unresolved PackageLocationImmutable))
+          repoObject value = do
+            pm <- parseJSON value
+            repo <- parseJSON value
+            pure $ noJSONWarnings $ pure $ PLIRepo repo pm
+
+          archiveObject value = do
+            pm <- parseJSON value
+            withObjectWarnings "UnresolvedPackageLocationImmutable.PLIArchive" (\o -> do
+              Unresolved mkArchiveLocation <- parseArchiveLocationObject o
+              archiveHash <- o ..: "sha256"
+              archiveSize <- o ..: "size"
+              archiveSubdir <- o ..:? "subdir" ..!= ""
+              pure $ Unresolved $ \mdir -> do
+                archiveLocation <- mkArchiveLocation mdir
+                pure $ PLIArchive Archive {..} pm
+              ) value
+
+          hackageObject value =
+             withObjectWarnings "UnresolvedPackagelocationimmutable.PLIHackage (Object)" (\o -> do
+                      treeKey <- o ..: "pantry-tree"
+                      htxt <- o ..: "hackage"
+                      case parseHackageText htxt of
+                        Left e -> fail $ show e
+                        Right (pkgIdentifier, blobKey) ->
+                          pure $ pure $ PLIHackage pkgIdentifier blobKey (TreeKey treeKey)) value
+
+          github value = do
+            pm <- parseJSON value
+            withObjectWarnings "UnresolvedPackagelocationimmutable.PLIArchive:github" (\o -> do
+              GitHubRepo ghRepo <- o ..: "github"
+              commit <- o ..: "commit"
+              let archiveLocation = ALUrl $ T.concat
+                    [ "https://github.com/"
+                    , ghRepo
+                    , "/archive/"
+                    , commit
+                    , ".tar.gz"
+                    ]
+              archiveHash <- o ..: "sha256"
+              archiveSize <- o ..: "size"
+              archiveSubdir <- o ..: "subdir"
+              pure $ pure $ PLIArchive Archive {..} pm) value
+
 instance FromJSON (WithJSONWarnings (Unresolved (NonEmpty RawPackageLocationImmutable))) where
   parseJSON v
       = http v
@@ -1497,7 +1591,7 @@ instance FromJSON (WithJSONWarnings (Unresolved (NonEmpty RawPackageLocationImmu
     <|> repo v
     <|> archiveObject v
     <|> github v
-    <|> fail ("Could not parse a UnresolvedPackageLocationImmutable from: " ++ show v)
+    <|> fail ("Could not parse a UnresolvedRawPackageLocationImmutable from: " ++ show v)
     where
       http :: Value -> Parser (WithJSONWarnings (Unresolved (NonEmpty RawPackageLocationImmutable)))
       http = withText "UnresolvedPackageLocationImmutable.UPLIArchive (Text)" $ \t ->
@@ -1910,6 +2004,27 @@ instance NFData SnapshotLocation
 instance ToJSON SnapshotLocation where
   toJSON sl = toJSON (toRawSL sl)
 
+instance FromJSON (WithJSONWarnings (Unresolved SnapshotLocation)) where
+    parseJSON v = file v <|> url v <|> compiler v
+      where
+        file = withObjectWarnings "SLFilepath" $ \o -> do
+           ufp <- o ..: "filepath"
+           pure $ Unresolved $ \mdir ->
+             case mdir of
+               Nothing -> throwIO $ InvalidFilePathSnapshot ufp
+               Just dir -> do
+                 absolute <- resolveFile dir (T.unpack ufp)
+                 let fp = ResolvedPath (RelFilePath ufp) absolute
+                 pure $ SLFilePath fp
+        url = withObjectWarnings "SLUrl" $ \o -> do
+          url' <- o ..: "url"
+          sha <- o ..: "sha256"
+          size <- o ..: "size"
+          pure $ Unresolved $ \_ -> pure $ SLUrl url' (BlobKey sha size)
+        compiler = withObjectWarnings "SLCompiler" $ \o -> do
+          c <- o ..: "compiler"
+          pure $ Unresolved $ \_ -> pure $ SLCompiler c
+
 -- | Convert snapshot location to its "raw" equivalent.
 --
 -- @since 0.1.0.0
@@ -2151,3 +2266,10 @@ getGlobalHintsFile = do
   root <- view $ pantryConfigL.to pcRootDir
   globalHintsRelFile <- parseRelFile "global-hints-cache.yaml"
   pure $ root </> globalHintsRelFile
+
+-- | Creates BlobKey for an input ByteString
+--
+-- @sinc 0.1.0.0
+bsToBlobKey :: ByteString -> BlobKey
+bsToBlobKey bs =
+    BlobKey (SHA256.hashBytes bs) (FileSize (fromIntegral (B.length bs)))